requirements.txt
adjustText==1.2.0
gdown==5.2.0
matplotlib==3.9.2
pandas==2.2.2
scikit_learn==1.5.1
seaborn==0.13.2
SQLAlchemy==2.0.32
streamlit==1.37.1
databricks-sql-connector[sqlalchemy]==3.4.0
python-dotenv==1.0.1